pdf-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0